+++ python/nano/src/bigdl/nano/deps/automl/hpo_api.py
@@ -31,3 +31,9 @@ def check_hpo_status(searcher):
                 dependency is not installed.', UserWarning)
         return False
     return True
+
+
+def create_optuna_backend():
+    """Create an Optuna Backend."""
+    from bigdl.nano.deps.automl.optuna_backend import OptunaBackend
+    return OptunaBackend()
